Sphere/transformer_roformer_files/state-tracking-transformer/.ipynb_checkpoints/requirements-checkpoint.txt
transformers>=4.38.0
torch>=2.0.0
pyyaml
transformers[torch]
